+++ CHANGELOG.md
@@ -5,3 +5,4 @@ For changes pre v1.5 see the [v1.4](https://github.com/absinthe-graphql/absinthe
 ## v1.5.0-dev
 
 - Complete rewrite of schema internals
+- Breaking change: added `node_name/0` callback to `Absinthe.Subscription.PubSub` behaviour. To retain old behaviour, implement this callback to return `Kernel.node/0`.